Changes the way that things are drawn. Now we have one stage whose opacity is set to 0.004 that registers all the mouse events. There is another stage who renders all these events as drawings. This should fix the issue with catching mouse events in the linux version.
As it currently works, the stage that renders the drawings does not yet follow the stage that picks up the listeners. The one that picks up the listeners can be moved around with the title bar (Although the bar is hard to see). This does lead to some bugs if you try and move the stage currently.

Another change that was added was a cosmetic change to the ControllerBox.
Now, the color picker is centered.